feat(tvix/eval): Support builtins.elemAt
(Attempt to) index into a list.

+++ tvix/eval/src/errors.rs
@@ -25,9 +25,9 @@ pub enum ErrorKind {
         name: String,
     },
 
-    // Attempted to index into a list beyond its boundaries.
+    /// Attempted to index into a list beyond its boundaries.
     IndexOutOfBounds {
-        index: usize,
+        index: i64,
     },
 
     TypeError {
